CLASSES-1606 Change poll bar graph orientation to horizontal and include alpha-counter on answers
CLASSES-1606 If a poll answer is marked as 'correct', then display this to the user after they have answered the question

CLASSES-1606 Change wording of message displayed to instructors when previewing a lessons poll

CLASSES-1606 Introduce new checkbox option for multiplechoice and shortanswer questions to show the correct answers to the student after response.  Also introduce a table view when creating a multiple choice question to help provide context for the checkbox (correct answer) and text field (answer text).

CLASSES-1606 Show letters for the correct answers if question type is multiplechoice.  Continue to show the answer text if question type is shortanswer.

CLASSES-1606 Bold the correct answer message