CLASSES-3420 Don't clear Quartz jobs on non-jobscheduler nodes
Otherwise our Attendance jobs get zapped whenever a server is
restarted.  This happens to attendance but not other jobs because we
added them to the DEFAULT Quartz group (so they would show up in the
Job Scheduler tool).  But since the Job Scheduler assumes jobs in that
group will be standard Sakai ones (registered with ComponentManager),
ours end up getting deleted.